Making Authorized Requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,21 @@
 import { useWorkoutContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 //date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const WorkoutDetails =({workout})=>{
-    
     const { dispatch } = useWorkoutContext();
+    const { user } = useAuthContext();
     const handleDelete = async()=>{
+        if(!user){
+            return
+        }
         const response = await fetch('/api/workouts/' +workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
 
         const json = await response.json();
